+++ languagetool-standalone/CHANGES.md
@@ -2,6 +2,10 @@
 
 ## 5.8-SNAPSHOT (release planned for 2022-06-28)
 
+#### Polish
+* updated the spelling dictionary to match changes in language (the current version is in sync with sjp.pl as of April 1, 2022)
+* added and improved rules
+
 ...